+++ api/views.py
@@ -112,21 +112,5 @@ def email_auth(request):
         return HttpResponseBadRequest("Некорректный метод запроса")
     
 
-def check_token(nickname, tokenVal):
-    user = UserAuth.objects.filter(nickname= nickname).get()
-    if user is None:
-        return False
-    
-    token = AuthTokens.objects.filter(user = user).filter(token = tokenVal).get() 
-    if token is None:
-        return False   
 
-    if token.type != "email":
-        return False
-    
-    if datetime.datetime.now().timestamp() > token.expiration_date.timestamp():
-        token.delete()
-        return False
-     
-    return True
