+++ home/templates/index.html
@@ -1,0 +1,3 @@
+{% extends "base.html" %}
+{% block title %}Index{% endblock title %}
+{% block content %}Главная{% endblock content %}
