+++ home/templates/login_base.html
@@ -1,0 +1,29 @@
+<!DOCTYPE html>
+<html>
+<head>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Skatert</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0-alpha2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-aFq/bzH65dt+w6FI2ooMVUpc+21e0SRygnTpmBvdBgSdnuTN7QbdgL+OapgHtvPp" crossorigin="anonymous">
+    {% block js %} 
+    {% endblock js %} 
+  </head>
+<body>
+    <h2>Login page</h2>
+    <div class="container">
+        <div class="row">
+          <div class="col">
+          </div>
+          <div class="col">
+          </div>
+          
+          <div class="col">
+            {% block content %}
+              
+            {% endblock content %}
+          </div>
+        </div>
+      </div>
+      
+</body>
+</html>
