+++ home/templates/user_page.html
@@ -1,0 +1,3 @@
+{% extends "base.html" %}
+{% block title %}{{user.nickname}} | Skatert {% endblock title %}
+{% block content %}Пользователь {{user.nickname}} {% endblock content %}
