+++ skatert/urls.py
@@ -18,7 +18,7 @@ from django.urls import include, path
 
 urlpatterns = [
     path('', include('home.urls')),
-    path('users/', include('users.urls')),
+    path('user/', include('users.urls')),
     path('api/', include('api.urls')),
     path('admin/', admin.site.urls),
 ]
